requirements.txt
# Your requirements go here
cachetools==4.2.2
google-auth==2.0.2
google-auth-oauthlib==0.4.6
gspread==4.0.1
oauthlib==3.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests-oauthlib==1.3.0
rsa==4.7.2